requirements/base.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.0
python-decouple==3.8
django-phonenumber-field[phonenumberslite]==8.0.0
phonenumberslite==8.13.40
rich==13.7.1
requests
tqdm==4.66.4
django-autocomplete-light==3.11.0
django-unfold==0.32.0
psycopg2-binary==2.9.9
django-environ==0.11.2
minio==7.2.7
whitenoise==6.7.0
Faker==26.0.0
pillow==10.4.0
nanoid==2.0.0
django-nanoid-field==0.1.4
django-redis==5.4.0
django-memoize==2.3.1
django-select2==8.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
